update ,ethods related to district

+++ LX.StaffScheduler.BLL/Services/Common/DistrictService.cs
@@ -34,16 +34,23 @@ namespace LX.StaffScheduler.BLL.Services.Common
             return district?.ToDTO();
         }
 
-        public async Task RemoveAsync(DistrictDTO entity)
+        public async Task RemoveAsync(int id)
         {
-            var district = entity.FromDTO();
-            repository.RemoveAsync(district);
+            var district = await repository.GetByIdAsync(id);
+            if (district != null)
+            {
+                await repository.RemoveAsync(district);
+            }
         }
 
         public async Task UpdateAsync(DistrictDTO entity)
         {
-            var district = entity.FromDTO();
-            await repository.UpdateAsync(district);
+            var district = await repository.GetByIdAsync(entity.Id);
+            if (district != null)
+            {
+                district.Name = entity.Name;
+                await repository.UpdateAsync(district);
+            }
         }
 
     }
